backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
motor==3.6.0
pymongo==4.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pypdf==5.1.0
faiss-cpu==1.9.0
sentence-transformers==3.3.1
numpy==2.1.3
openai==1.54.4
email-validator==2.2.0
redis==5.1.1
